Adjacent faces and Intrinsic data term
MeshData now includes also the list of adjacent faces per vertex. This
is set in MeshIO when loading the mesh from file.
New photometric cost defined as the difference between the estimated
intensity (known albedo * estimated shading as a function of spherical
harmonic coefficients and normal direction) and the intensity of the
projected 3d point in the image.
Normals are computed as a the average of face normals using the saved
list of adjacent faces.