Intrinsic AutoDiffCost and solved bug of () operator
Since the AutoDiffCostFunctions for the case of RGB photometric error
was causing problems, now only DynamicAutoDiff is used as a workaround.
It should be checked if this is slower and, if so, solved it in the
future.
The () operator of the Intrinsic residual was not computing the normal
direction properly since the neighbour vertices were defined locally and
released before been used.