Removed has_full_ring and smoothing term added
The attribute has_full_ring from MeshData has been removed. Since we
have the ordered adjacent vertices and faces, we know that a vertex has
a closed one-ring if the number of adjacent faces is equal to the number
of adjacent vertices.
The smoothing regularisation term has been added. The label in the
configuration file is 'smoothing_weight'